requirements.txt
django
django-cors-headers
django-templated-mail
djangorestframework
djoser
pyyaml
uritemplate
